requirements.txt
<<<<<<< HEAD
Django==4.2.7
djangorestframework==3.14.0
redis==5.0.1
django-redis==5.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9  # If using PostgreSQL
whitenoise==6.6.0  # For static files in production
django-ckeditor
django-cors-headers
django-ckeditor-5
=======
django
djangorestframework
django-ckeditor
googletrans==4.0.0-rc1

>>>>>>> cf9f1cb (feat: Add multilingual FAQ model with Redis caching)
